backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
asyncpg  
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
greenlet
bcrypt==4.0.1

